refactor(worker): 로거 내부 클래스로 이동
